Modularise PDF manipulation tools
