+++ app/models/concerns/securable.rb
@@ -24,10 +24,11 @@ module Securable
     return if secured? || content.blank? || !pdf?
     return if (images = convert_original_content_to_images(density)).empty?
 
-    secured = convert_images_to_pdf(images)
-    update!(secured_content: secured)
-    secured.close
-    delete_files([secured.path])
+    secured_filename = content.filename
+    PdfUtils.convert_images_to_pdf(images, secured_filename)
+    update!(secured_content: File.open(secured_filename))
+
+    PdfUtils.delete_files([secured_filename] + images)
   end
 
   private
@@ -35,9 +36,9 @@ module Securable
   def convert_original_content_to_images(density)
     original_filename = content.filename.parameterize
     original_file = download(original_filename, content.read)
-    image_filenames = convert_to_images(original_file, density)
+    image_filenames = PdfUtils.convert_pdf_file_to_images(original_file, density)
     original_file.close
-    delete_files([original_filename])
+    PdfUtils.delete_files([original_filename])
     image_filenames
   end
 
@@ -49,38 +50,6 @@ module Securable
     file
   end
 
-  def convert_to_images(pdf_file, density)
-    image = MiniMagick::Image.open(pdf_file.path)
-    image.pages.each_with_index { |page, index| convert_page_with_mini_magick(page.path, index, density) }
-    Dir.entries(".").select { _1.start_with?("secured-image-#{id}-") }.sort
-  rescue # ImageMagick failure: sometimes the pdf can't be opened
-    []
-  end
-
-  def convert_page_with_mini_magick(page_path, index, density)
-    MiniMagick::Tool::Convert.new do |convert|
-      convert << "-quality" << "100"
-      convert << "-density" << density.to_s
-      convert << page_path
-      convert << "secured-image-#{id}-#{index.to_s.rjust(5, "0")}.jpg"
-    end
-  rescue # ImageMagick failure: sometimes the page is not convertable
-    nil
-  end
-
-  def convert_images_to_pdf(image_filenames)
-    secured_filename = content.filename
-    MiniMagick::Tool::Convert.new do |convert|
-      image_filenames.each { convert << _1 }
-      convert << secured_filename
-    end
-    delete_files(image_filenames)
-    File.open(secured_filename)
-  end
-
-  def delete_files(filenames)
-    filenames.select { File.exist?(_1) }.each { File.delete(_1) }
-  end
 
   def pdf?
     content.content_type == "application/pdf"
